+++ test/index.js
@@ -9,8 +9,14 @@ describe('gracenode initialization ->', function () {
 		gn.setConfigPath('node_modules/gracenode/test/configs/');
 		gn.setConfigFiles(['index.json']);
 
+		gn.addModulePath('node_modules/gracenode/test/modules/');
+
+		// test override
+		gn.override('view');
+
 		gn.setup(function (error) {
 			assert.equal(error, undefined);
+			assert.equal(gn.view.test, 'test override');
 			done();
 		});