+++ test/modules/view/index.js
@@ -1,0 +1,13 @@
+var gn = require('../../../');
+var logger = gn.log.create('overridden-view');
+
+module.exports.readConfig = function () {
+	logger.info('overriden view.readConfig');
+};
+
+module.exports.setup = function (cb) {
+	logger.info('overridden view.setup');
+	cb();
+};
+
+module.exports.test = 'test override';